Refactor side_load into a separate  class
